+++ neue.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-    css: {
-        extract: true
-    }
+  css: {
+    extract: true
+  }
 };
