--- ssr/server-bundle-test.js
@@ -1,12 +1,0 @@
-// =================
-// Usage
-// =================
-require('../dist/js/server-bundle.js')
-
-// var promise = global.renderServer()
-//
-// promise.then(value => {
-//     console.log(value); // Success!
-// }, reason => {
-//     console.log(reason); // Error!
-// });
